sim: register external signal handler in initSignals
This commit adds initSigRtmin() to initSignals() to try to
register the external signal handler earlier in gem5 execution.
This change was part of an unsuccessful attempt to get the
dashboard to update with new progress bars if new gem5
simulations were launched. When the gem5 dashboard checked
for newly launched gem5 simulations, it was causing the new
simulations to exit by sending SIGHUP (to get stats)
before the handler was registered. Hence, the function to
register the external signal handler was added to a function that
is called earlier.