sim: fix signal handler for hypercall external signal
This commit changes the signal handler for external gem5
orchestrator signals, which uses SIGHUP, so that it matches the
signal handlers for other signals. It was previously set up
differently/incorrectly, which could have been the cause for
seg faults that occurred when simulations were connected to the
gem5 dashboard.