sim, util: change SIGHUP to SIGCONT for shared mem transmitter
This commit changes the signal for the shared memory transmitter,
used by the gem5 orchestrator and dashboard, from SIGHUP to
SIGCONT. If no signal handlers are registered, SIGHUP terminates
processes and SIGCONT is ignored. This change fixes the issue of
the gem5 dashboard killing newly launched gem5 simulations if
they receive SIGHUP before the signal handlers are registered.